[refactor] 전체적인 오류 수정 및 레이아웃 수정 #118
