--- client/js/common/index.js
@@ -1,1 +1,0 @@
-export * from './main-banner.js';
