+++ client/js/main/index.js
@@ -3,3 +3,5 @@ export * from './recently-viewed.js';
 export * from './recently-item.js';
 export * from './recently-swiper.js';
 export * from './main-product-list-item.js';
+export * from './main-banner.js';
+export * from './main-banner-slide.js';
