+++ client/register.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8" />
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  
+  <link rel="icon" type="image/png" sizes="16x16" href="./assets/icons/Icon/favicon-16x16.png">
   <link rel="stylesheet"
   href="https://cdn.jsdelivr.net/npm/swiper@8/swiper-bundle.min.css" />
   <link rel="stylesheet" href="./css/main.css" />
@@ -280,7 +280,7 @@
                 <label for="terms-of-event">
                   <div class="check-n-title">
                     <input type="checkbox" id="terms-of-event" name="check-event" class="check-event__input" value="" />
-                    <span class="check-event__title check__title">무료배송,할인쿠폰 등 혜택/정보 수신 동의 (선택)</span>
+                    <span class="check-event__title check__title">무료배송,할인쿠폰 등 혜택/정보수신동의(선택)</span>
                   </div>
                   <a href="">약관보기 <span class="check__arrow"></span></a>
                 </label>
