+++ client/sass/layout/_navigation.scss
@@ -4,7 +4,7 @@
 
 
 .navigation {
-  
+  border-bottom: 1px solid $gray-200;
   &-container {
     display: flex;
     justify-content:space-between;
@@ -13,7 +13,6 @@
     width: rem(1050px);
     height: rem(72px);
     margin: 0 auto;
-    border-bottom: 2px solid $gray-200;
     font-family: "Pretendard Variable", Pretendard, -apple-system, BlinkMacSystemFont, system-ui, Roboto, "Helvetica Neue", "Segoe UI", "Apple SD Gothic Neo", "Noto Sans KR", "Malgun Gothic", "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", sans-serif;
   }
 
@@ -65,7 +64,7 @@
     display: none;
     position: absolute;
     top: rem(72px);
-    z-index: 10;
+    z-index: 20;
     background-color: $white;
 
     &__list {
